Use `npx` for Mintlify commands
There is a bug that causes `mintlify dev` to use a very old version of
Mintlify when the `mintlify` npm package is installed as a dependency,
even when running the command via `npx`.  For example:

  ```console
  $ npm exec mintlify version
  cli version 0.1.0
  client version 0.0.1942

  $ npm uninstall mintlify && npm install --save-dev mintlify

  $ git diff package.json
  diff --git a/package.json b/package.json
  index e6906f7..1472064 100644
  --- a/package.json
  +++ b/package.json
  @@ -40,7 +40,7 @@
       "ajv": "^8.17.1",
       "ajv-formats": "^3.0.1",
       "glob": "^11.0.0",
  -    "mintlify": "^4.0",
  +    "mintlify": "^4.2.183",
       "prettier": "^3.6.2",
       "tsx": "^4.19.1",
       "typedoc": "^0.28.7",

  $ npm exec mintlify version
  cli version 0.1.0
  client version 0.0.1942

  $ npx mintlify@4.2.183 version
  cli version 0.1.0
  client version 0.0.1942

  $ npm uninstall mintlify

  $ npx mintlify version
  cli version 4.2.183
  client version 0.0.1942
  ```

This commit removes `mintlify` as a dev dependency and changes
Mintlify-related commands (e.g., `npm run serve:docs`), to use
`npx mint` (e.g., `cd docs && npx mint dev`).  (`mint` is now the
preferred npm package.  It, unfortunately, exhibits the same bug as the
`mintlify` package.)

This fixes many of the discrepancies that we have been experiencing
between the local preview and the production site.